// next-shell/src/types/term_form.types.ts
export interface TermFormData {
    supplier: string;
    supplierName: string;
    contactPerson: string;
    contactPersonName: string;
    contractNo: string;
    mfgPartNo: string;
    suppOrderCode: string;
    active: boolean;
    validFrom: string;
    validTo: string;
    purchaseTerm: string;
    purchaseTermLocation: string;
    purchaseTermLocationName: string;
    purchaseSubLocation: string;
    salesTerm: string;
    salesSubLocation: string;
    incoterm: string;
    prodCost: number;
    pkh: number;
    soc: number;
    currency: string;
    exRate: number;
    shipMode: string;
    dimUnit: string;
    length: number;
    width: number;
    height: number;
    weight: number;
    cWeight: number;
    freightType: string;
    freightRate: number;
    fr: number;
    insPercent: number;
    dutyPercent: number;
    excisePercent: number;
    miscTax: number;
    numInBuy: number;
    numInSale: number;
    markup: number;
    stockFeePercent: number;
    wireTT: number;
    customClear: number;
    scc: number;
    zoneRate: number;
    salesPerson: string;
    salesPersonName: string;
    sourcedBy: string;
    sourcedByName: string;
    remark: string;
    leadTime: string;
    moq: string;
    vendorBPA: boolean;
    purchaseUOM: string;
    salesUOM: string;
    spk: number;
    qoc: number;
    updatedBy: string;
    updatedDate: string;
    stockUOM: string;
}
// ...
export interface TermCalcResults {
    OP1: number;
    OP1_THB: number;
    OP2_THB: number;
    DIM_WEIGHT: number;
    SHP_WEIGHT: number;
    INS: number;
    FR_QTEC: number;
    FR_ZONE: number;
    CIF: number;
    CIF_ZONE: number;
    DT: number;
    DT_FR: number;
    DT_ZONE: number;
    ET: number;
    MT: number;
    PRE_QLC: number;
    STK: number;
    QLC: number;
    QLC2: number;
    QLC3: number;
    TOTAL_PRICE: number;
    MK_THB: number;
    SALES_PRICE: number;
}
// ...
export const defaultTermFormData: TermFormData = {
    supplier: '',
    supplierName: '',
    contactPerson: '',
    contactPersonName: '',
    contractNo: '',
    mfgPartNo: '',
    suppOrderCode: '',
    active: true,
    validFrom: '',
    validTo: '',
    purchaseTerm: '',
    purchaseTermLocation: '',
    purchaseTermLocationName: '',
    purchaseSubLocation: '',
    salesTerm: '',
    salesSubLocation: '',
    incoterm: 'Incoterms 2020',
    prodCost: 0,
    pkh: 0,
    soc: 0,
    currency: '',
    exRate: 1,
    shipMode: '-1',
    dimUnit: '1',
    length: 0,
    width: 0,
    height: 0,
    weight: 0,
    cWeight: 0,
    freightType: '',
    freightRate: 0,
    fr: 0,
    insPercent: 1,
    dutyPercent: 0,
    excisePercent: 0,
    miscTax: 0,
    numInBuy: 1,
    numInSale: 1,
    markup: 0,
    stockFeePercent: 0,
    wireTT: 0,
    customClear: 0,
    scc: 0,
    zoneRate: 0,
    salesPerson: '',
    salesPersonName: '',
    sourcedBy: '',
    sourcedByName: '',
    remark: '',
    leadTime: '',
    moq: '',
    vendorBPA: false,
    purchaseUOM: '',
    salesUOM: '',
    spk: 0,
    qoc: 0,
    updatedBy: '',
    updatedDate: '',
    stockUOM: 'EA',
};